+++ app/src/features/incentivization/hooks/useIsIncentivizationEnabled.ts
@@ -1,50 +1,15 @@
 import { useEffect, useState } from "react";
-import { useSelector } from "react-redux";
 import { useFeatureValue } from "@growthbook/growthbook-react";
-import {
-  getDaysSinceSignup,
-  getExtensionInstallDate,
-  getExtensionSignupDate,
-  getUserAuthDetails,
-} from "store/selectors";
-import moment from "moment";
+import { useIsNewUserForIncentivization } from "./useIsNewUserForIncentivization";
 
 export const useIsIncentivizationEnabled = () => {
-  const user = useSelector(getUserAuthDetails);
-  const extensionInstallDate = useSelector(getExtensionInstallDate);
-  const extensionSignupDate = useSelector(getExtensionSignupDate);
-  const daysSinceSignup = useSelector(getDaysSinceSignup) ?? 0;
   const [isEnabled, setIsEnabled] = useState(false);
-
   const isFeatureFlagEnabled = useFeatureValue("incentivization_onboarding", false);
+  const isNewUser = useIsNewUserForIncentivization("2024-06-20");
 
   useEffect(() => {
     const getIsIncentivizationEnabled = () => {
-      const releaseDate = new Date("2024-06-20").getTime();
-      const currentDate = new Date().getTime();
-
-      if (currentDate < releaseDate) {
-        return false;
-      }
-
-      if (isFeatureFlagEnabled) {
-        if (!user?.loggedIn) {
-          return extensionInstallDate && new Date(extensionInstallDate).getTime() >= releaseDate;
-        } else {
-          if (extensionSignupDate) {
-            if (new Date(extensionSignupDate).getTime() >= releaseDate) {
-              return !extensionInstallDate || new Date(extensionInstallDate).getTime() >= releaseDate;
-            }
-          } else {
-            // TODO: add sentry log for this case
-            const momentDate = moment(currentDate);
-            const signupDate = momentDate.subtract(daysSinceSignup, "days");
-            return signupDate.toDate().getTime() >= releaseDate;
-          }
-        }
-      }
-
-      return false;
+      return isFeatureFlagEnabled && isNewUser;
     };
 
     const timerId = setTimeout(() => {
@@ -57,7 +22,7 @@ export const useIsIncentivizationEnabled = () => {
         clearTimeout(timerId);
       }
     };
-  }, [isFeatureFlagEnabled, user?.loggedIn, extensionInstallDate, extensionSignupDate, daysSinceSignup]);
+  }, [isFeatureFlagEnabled, isNewUser]);
 
   useEffect(() => {
     window.incentivization = isEnabled;
